assorted changes

+++ include/cache.hpp
@@ -12,7 +12,7 @@ class UserCache {
 public:
     std::string access_token =  "";
     std::string refresh_token = "";
-    uint32_t expires_at =       0;
+    uint64_t expires_at =       0;
 
     UserCache() = default;
     UserCache(toml::value &value);
@@ -48,7 +48,7 @@ public:
     void UpdateUserCache(
         std::string access_token,
         std::string refresh_token,
-        uint32_t    expires_at
+        uint64_t    expires_at
     );
     void UpdatePlaybackCache(
         std::string spotify_id,
